+++ src/engine/Rendering/Renderer.hxx
@@ -14,8 +14,8 @@ enum RendererBackendType : uint8_t {
 
 auto RendererInit(
 	SDL_Window* window,
-	uint16_t width,
-	uint16_t height,
+	uint32_t width,
+	uint32_t height,
 	RendererBackendType backend
 ) -> uint8_t;
 auto RendererUpdate() -> void;
